backend/requirements.txt
torch
torchvision
transformers
fastapi
uvicorn
numpy
pillow
clip-by-openai